give up recognizing comments
It is so hard to recognize comments in a source code file. I'm afraid
that we will fail to finish the necessary functions before the deadline.
Meanwhile I added a button to skip the current character if it is unable
to be typed correctly. Thanks to the advises given by my professor.
Happy Chinese new year to all of you.